Reduced warnings.
git-svn-id: https://svn.aros.org/svn/aros/trunk@19685 fb15a70f-31f2-0310-bbcc-cdcc74a49acc

+++ workbench/libs/mathtrans/spcos.c
@@ -59,7 +59,7 @@ AROS_LH1(float, SPCos,
     LONG z,Res,ysquared,yabs,tmp;
     yabs = fnum1 & (FFPMantisse_Mask + FFPExponent_Mask);
     
-    if (FFP_Pinfty == yabs)
+    if ((LONG)FFP_Pinfty == yabs)
     {
         SetSR(Overflow_Bit, Zero_Bit | Negative_Bit | Overflow_Bit);
         return FFP_NAN;
